+++ tvsd/actions.py
@@ -3,31 +3,36 @@ import logging
 import os
 import sys
 from typing import List, Tuple
-from tvsd._variables import state_base_path, state_temp_base_path, state_series_dir
-from tvsd import state
-
-from tvsd.download import Download
-from tvsd.search import SearchQuery
-from tvsd.utils import check_dir_mounted, is_video
 
 from rich.console import Console
 from rich.table import Table
+
+from tvsd._variables import state_base_path, state_series_dir, state_temp_base_path
+from tvsd.download import Download
+from tvsd.search import SearchQuery
+from tvsd.utils import check_dir_mounted, is_video
 
 
 def search_media_and_download(query: str, specials_only: bool = False):
     """Search for media and download
 
+    This function searches for media based on the given query string and downloads it.
+    It first checks if the base path is mounted and exits if it is not.
+    Then it searches for the given query and finds the show.
+    Finally, it starts a guided download of the chosen show.
+
     Args:
         query (str): query string
+        specials_only (bool): Download only specials episode. Defaults to False.
     """
-    logging.info(f"Checking if {state_base_path()} is mounted...")
+    logging.info("Checking if %s is mounted...", state_base_path())
     if not check_dir_mounted(path=state_base_path()):
         sys.exit()
     logging.debug("Base path: %s", state_base_path())
 
     # Search
     query_instance = SearchQuery(query)
-    logging.info(f"Searching for {query}...")
+    logging.info("Searching for %s...", query)
     query_instance.find_show(state_base_path())
 
     # Download
@@ -37,7 +42,7 @@ def search_media_and_download(query: str, specials_only: bool = False):
         temp_path=state_temp_base_path(),
         specials_only=specials_only,
     )
-    logging.info(f"Starting {query_instance.chosen_show.title} guided download...")
+    logging.info("Starting %s guided download...", query_instance.chosen_show.title)
     download_instance.guided_download()
 
 
@@ -50,10 +55,10 @@ def list_shows_as_table(show_index=False) -> Tuple[List[str], int]:
     Returns:
         Tuple[List[str], int]: List of shows and number of shows
     """
-    dir = os.path.join(state_base_path(), state_series_dir())
-    logging.info(f"Checking if {dir} exists...")
-    if not os.path.isdir(dir):
-        logging.error(f"{dir} does not exist! Nothing to list. Exiting...")
+    series_dir = os.path.join(state_base_path(), state_series_dir())
+    logging.info("Checking if %s exists...", series_dir)
+    if not os.path.isdir(series_dir):
+        logging.error("%s does not exist! Nothing to list. Exiting...", series_dir)
         sys.exit()
 
     console = Console()