+++ tvsd/types/episode.py
@@ -1,63 +1,103 @@
+"""
+Episode class.
+"""
+
+import logging
 import os
 import re
 from typing import TYPE_CHECKING
 
+from tvsd._variables import state_base_path
 from tvsd.utils import file_exists_in_base, relative_to_absolute_path
 
-
 if TYPE_CHECKING:
-    from tvsd.season import Season
+    from tvsd.types.season import Season
 
 
 class Episode:
-    """Episode class"""
+    """A class representing an episode of a TV show.
+
+    Attributes:
+        _name (str): The name of the episode.
+        _url (str): The URL of the episode.
+        _number (int): The episode number.
+        _season (Season): The season that the episode belongs to.
+        _not_specials (bool): True if the episode is not a special episode, False otherwise.
+    """
 
     def __init__(
         self,
         episode_name: str,
         episode_url: str,
-        # previous_episode: "Episode" = None,
-        # next_episode: "Episode" = None,
         season: "Season",
     ):
+        """Initialize a new Episode object.
+
+        Args:
+            episode_name (str): The name of the episode.
+            episode_url (str): The URL of the episode.
+            season (Season): The season that this episode belongs to.
+        """
         self._name = episode_name
         self._url = episode_url
         self._number: int
 
-        # self._previous_episode = previous_episode
-        # self._next_episode = next_episode
         self._season: "Season" = season
 
         self._not_specials = True
 
-    def __str__(self):
+    def __str__(self) -> str:
+        """
+        Returns a string representation of the Episode object.
+
+        The string representation includes the name and episode number of the episode.
+
+        Returns:
+            str: A string representation of the Episode object.
+        """
         return f"{self.name} ({self.episode_number})"
 
     def __repr__(self):
+        """
+        Returns a string representation of the Episode object.
+        The string contains the name and episode number of the episode.
+        """
         return f"{self.name} ({self.episode_number})"
 
     def identify_episode_number_from_name(self) -> int:
-        """Tries to identify the episode number from the episode name.
+        """
+        Tries to identify the episode number from the episode name.
+
+        This function uses a regular expression to identify the episode number from the episode name.
+        If the episode number is found, it is returned as an integer. If it is not found, the function
+        returns 1.
 
         Returns:
             int: The identified episode number.
         """
-
         try:
             print(self._name)
             episode_number_identifying_regex = r"^[0-9]{8}[（(]*第(\d+)[期集][(（上中下)）]*[)）]?$|^(\d{1,3})$|^第(\d+)[期集][上中下]*$"
             episode_number_match = re.match(
                 episode_number_identifying_regex, self._name
-            ).groups()
+            )
 
-            # Filter out all None matches
-            identified_number = [i for i in episode_number_match if i is not None][0]
+            if episode_number_match is not None:
+                episode_number_groups = episode_number_match.groups()
 
-            print(f"episode_number: {identified_number}")
-            # episode_index = int(re.findall(r'\d*', episode_number)[0])
-            resulting_index = int(re.findall(r"^\d{1,3}$", identified_number)[0])
+                # Filter out all None matches
+                identified_number = [i for i in episode_number_groups if i is not None][
+                    0
+                ]
 
-            # print(episode_number)
+                print(f"episode_number: {identified_number}")
+                # episode_index = int(re.findall(r'\d*', episode_number)[0])
+                resulting_index = int(re.findall(r"^\d{1,3}$", identified_number)[0])
+
+                # print(episode_number)
+            else:
+                resulting_index = 1
+
         except AttributeError:
             resulting_index = 1
 
@@ -98,6 +138,8 @@ class Episode:
     def episode_number(self) -> int:
         """Returns the episode number.
 
+        If the episode number is not already set, it will be identified from the episode name.
+
         Returns:
             int: The episode number.
         """
@@ -107,12 +149,19 @@ class Episode:
 
     @episode_number.setter
     def episode_number(self, episode_number: int):
-        """Sets the episode number."""
+        """
+        Sets the episode number.
+
+        Args:
+            episode_number (int): The episode number to set.
+        """
         self._number = episode_number
 
     @property
     def determine_episode_number(self) -> int:
         """Determines the episode number.
+
+        This method determines the episode number by either incrementing the previous episode number by 1 or by identifying the episode number from the episode name. The episode number is then stored in the instance variable _number.
 
         Returns:
             int: The episode number.
@@ -129,6 +178,8 @@ class Episode:
     def name(self) -> str:
         """Returns the episode name.
 
+        This method returns the name of the episode.
+
         Returns:
             str: The episode name.
         """
@@ -143,6 +194,9 @@ class Episode:
     def filename(self) -> str:
         """Returns the filename of the episode.
 
+        The filename is formatted as follows:
+        {show_prefix} - S{season_index}E{episode_number} - {episode_name}
+
         Returns:
             str: The filename of the episode.
         """
@@ -155,8 +209,7 @@ class Episode:
     def get_episode_url(self) -> str:
         """Gets the episode url from the episode object.
 
-        Args:
-            episode (Episode): The episode object.
+        This method returns the url of the episode object.
 
         Returns:
             str: The episode url.
@@ -166,7 +219,10 @@ class Episode:
 
     @property
     def relative_episode_file_path(self) -> str:
-        """Returns the relative path to the episode file.
+        """
+        Returns the relative path to the episode file.
+
+        The returned path is relative to the root directory of the TVSD project.
 
         Returns:
             str: The relative path to the episode file.
@@ -174,28 +230,42 @@ class Episode:
         return f"{self._season.relative_season_dir}/{self.filename}.mp4"
 
     @property
-    def file_exists_locally(self) -> bool:
-        """Returns True if the episode exists locally, False otherwise.
+    def file_exists_locally(self) -> str:
+        """Returns the name of the existing file if the episode exists locally, or an empty string otherwise.
 
         Returns:
-            bool: True if the episode exists locally, False otherwise.
+            filename(str): Name of existing file if the episode exists locally, Empty String otherwise.
         """
         # Check if file exists already
         if file_exists_in_base(self.relative_episode_file_path):
             print(f"{self.filename} already exists in directory, skipping... ")
-            return True
+            return self.filename
+
+        episode_title = self.filename.split(" - ")[-1]
+        try:
+            for file in os.listdir(
+                os.path.join(state_base_path(), self._season.relative_season_dir)
+            ):
+                if episode_title in file:
+                    print(f"{self.filename} probably exist as {file}, skipping...")
+                    return file
+        except FileNotFoundError:
+            logging.debug(
+                "%s does not exist, skipping file check in this directory...",
+                self._season.relative_season_dir,
+            )
 
         # specials exists already
         for existing_episode in os.listdir(
             relative_to_absolute_path(self.relative_destination_dir)
         ):
             # print(episode_name, existing_episode)
-            if existing_episode.endswith(".mp4") and self.filename in existing_episode:
+            if existing_episode.endswith(".mp4") and episode_title in existing_episode:
                 print(
                     f"{self.filename} probably exist as {existing_episode}, skipping..."
                 )
-                return True
-        return False
+                return existing_episode
+        return ""
 
     @property
     def season(self) -> "Season":
@@ -208,16 +278,27 @@ class Episode:
 
     @property
     def fetch_episode_m3u8_url(self) -> str:
-        """Fetches the m3u8 url of the episode.
+        """
+        Fetches the m3u8 url of the episode.
+
+        This method fetches the m3u8 url of the episode by calling the fetch_episode_m3u8 method of the season's source object,
+        passing in the relative episode url as a parameter. If the m3u8 url is not found, an empty string is returned.
 
         Returns:
             str: The m3u8 url of the episode.
         """
-        return self.season.source.fetch_episode_m3u8(relative_episode_url=self._url)
+        m3u8_url = self.season.source.fetch_episode_m3u8(relative_episode_url=self._url)
+        if m3u8_url is None:
+            return ""
+        return m3u8_url
 
     @property
     def relative_destination_dir(self) -> str:
         """Returns the relative destination directory of the episode.
+
+        If the episode is a special episode, the relative destination directory
+        will be the relative specials directory of the season. Otherwise, it will
+        be the relative season directory of the season.
 
         Returns:
             str: The relative destination directory of the episode.